+++ stingray/gti.py
@@ -4,11 +4,11 @@ import warnings
 from collections.abc import Iterable
 import copy
 
-from astropy.io import fits
 from .utils import contiguous_regions, jit, HAS_NUMBA
 from .utils import assign_value_if_none, apply_function_if_none
 from .utils import check_iterables_close, is_sorted
-from .io import fits_open_including_remote
+from .utils import fits_open_including_remote
+
 from stingray.exceptions import StingrayError
 from stingray.loggingconfig import setup_logger
